more tries

+++ Pandaros.Settlers/Pandaros.Settlers/Items/MeshRotationEuler.cs
@@ -8,8 +8,8 @@ namespace Pandaros.Settlers.Items
 {
     public class MeshRotationEuler
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public int x { get; set; }
+        public int y { get; set; }
+        public int z { get; set; }
     }
 }
